Issue #109
Fixes issue #109 where nested parentheses within the parameter signature additional traits caused a failure to correctly recognise the signature as valid due to not matching the ParameterAbbrevDefinitionRegex.
I have removed the regex in favour of manually parsing the parameter signature so as to make the code clearer and more maintainable than what the regex would have had to become in order to match nested brackets within the additional traits as well as possible additional brackets in the signature description.

Note: This was commit was orginally missed out from the shared/refactor-integration, pressumably by a mistake

Conflicts:
	src/ParameterDefinitonParser.h