Remove ResourceMethodSectionType. Fix separation of section types concerns.
ResourceMethodSectionType is remnant of the past and not needed under the current parser architecture. Similarly a `SectionProcessor` should never use another SectionsProcessor's regexes, they should be considered **private** to the respective processor. Instead expose a method that wraps the need for a processor's private refex. This way we will achieve separation of concerns.