Report ignored list item nodes. Fixes #163. Fixes #189.
There was a problem that any and all markdown nodes not handled at least by the `processNestedSection()` or
`isUnexpectedNode()` were silently ignored. This is because of the behavior of the "list adapter" where advancing
to a next node causes ignoring the rest of list item's nodes. This commit changes the behavior in a way that
every node (of a list item section) is treaded as unexpected if not handled by the `processNestedSection()`.