Treat subsequent nested sections as unexpected. Fixes #188.
Unexpected blocks that are recognized by subsequent nested parser should but aren't
immediate children of the actual section should be treated as unexpected sections in the
context of the actual parser. Therefore they will be handled as unexpected block by the actual
parser without giving back control to parents. Therefore the parser can recover from encountering
an alien block and continue to parse in the actual context.