Unify processor interface. Check boundaries while processing. Fixes #205
This commit unifies the interface of all `SectionProcessorBase` `process...` methods.
Previously, it was expected that `processSignature()`, `processDescription()` and `processContent()`
always processes exactly one node (or none). However with the Blueprint `SectionProcessor` processing (arguably not correctly)
multiple nodes it is clear that is essential to expect any `process...` methods to process an arbitrary number of nodes. For this
the interface of aforementioned nodes needed to be extended to allow checking the boundary of the node iterator.