Source map export is no longer the default for tests.
By an error the export of source map was turned on as default for all the test cases using `SectionParserHelper::parse`. This was incorrect as the export of source map is not the default behavior of the parser. Therefore most of the existing unit test were run with this option. As the result a code that is not executed during normal parsing was being executed during the tests making the tests not testing the default code path used in normal operating scenarios.

This commit changes the default behavior of `SectionParserHelper::parse()` to NOT use the `ExportSourcemapOption` option. This change alone is not enough to address the issue described above as most of the existing test are still executing code that is not normally executed. For that a much bigger change is needed – to separate source map related unit tests for other unit test.

The coupling of unit tests cases clearly points to the issue in the current test code base where many of the tests are not testing one particular unit (case). Therefore the tests are not relying on other test covering their respective test cases. This results in duplicate tests, more code, higher complexity and thus worse maintenance.